+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2025-06-24
+### Removed
+- Innecessary Embedded Apps
+
+### Fixed
+- User name
+- Language routing
+- Language bug while login and logout
+
+### Changed
+- Embed external libraries
+- Angular 18
+- Angular 19
+- Translations
+
+### Added
+- Teiler Backend config for Logo Text and Logo Height
+- Deactivate OIDC if OIDC-URL not set
+- Styling
+
+
 ## [1.0.4] - 2025-03-04
 ### Added
 - Assets API service
@@ -12,7 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.2] - 2024-02-14
 ### Changed
 - OIDC instead of KEYCLOAK
-
 
 ## [1.0.1] - 2023-12-13
 ### Fixed